+++ src/transformers/modeling_utils.py
@@ -53,7 +53,7 @@ from .pytorch_utils import (  # noqa: F401
     prune_layer,
     prune_linear_layer,
 )
-from .quantizers import AutoHFQuantizer, HFQuantizer
+from .quantizers import AutoHfQuantizer, HfQuantizer
 from .safetensors_conversion import auto_conversion
 from .utils import (
     ADAPTER_SAFE_WEIGHTS_NAME,
@@ -1056,7 +1056,7 @@ class ModuleUtilsMixin:
             total_parameters = list(self.parameters())
 
         total_numel = []
-        is_loaded_in_4bit = self.quant_method is not None and self.quant_method == "bitsandbytes_4bit"
+        is_loaded_in_4bit = getattr(self, "is_loaded_in_4bit", False)
 
         if is_loaded_in_4bit:
             if is_bitsandbytes_available():
@@ -1124,16 +1124,6 @@ class ModuleUtilsMixin:
         """
 
         return 6 * self.estimate_tokens(input_dict) * self.num_parameters(exclude_embeddings=exclude_embeddings)
-
-    @property
-    def quant_method(self) -> Optional[str]:
-        """
-        Optionally return the quantization type of the model. Returns None if the
-        model has not been quantized
-        """
-        hf_quantizer = getattr(self, "hf_quantizer", None) is not None
-        if isinstance(hf_quantizer, HFQuantizer):
-            return hf_quantizer.quantization_config.quant_method
 
 
 class PreTrainedModel(nn.Module, ModuleUtilsMixin, GenerationMixin, PushToHubMixin, PeftAdapterMixin):
@@ -2262,7 +2252,7 @@ class PreTrainedModel(nn.Module, ModuleUtilsMixin, GenerationMixin, PushToHubMix
 
         hf_quantizer = getattr(self, "hf_quantizer", None)
         quantization_serializable = (
-            hf_quantizer is not None and isinstance(hf_quantizer, HFQuantizer) and hf_quantizer.is_serializable
+            hf_quantizer is not None and isinstance(hf_quantizer, HfQuantizer) and hf_quantizer.is_serializable
         )
 
         if hf_quantizer is not None and not _hf_peft_config_loaded and not quantization_serializable:
@@ -3020,12 +3010,12 @@ class PreTrainedModel(nn.Module, ModuleUtilsMixin, GenerationMixin, PushToHubMix
         pre_quantized = getattr(config, "quantization_config", None) is not None
         if pre_quantized or quantization_config is not None:
             if pre_quantized:
-                config.quantization_config = AutoHFQuantizer.merge_quantization_configs(
+                config.quantization_config = AutoHfQuantizer.merge_quantization_configs(
                     config.quantization_config, quantization_config
                 )
             else:
                 config.quantization_config = quantization_config
-            hf_quantizer = AutoHFQuantizer.from_config(config.quantization_config, pre_quantized=pre_quantized)
+            hf_quantizer = AutoHfQuantizer.from_config(config.quantization_config, pre_quantized=pre_quantized)
         else:
             hf_quantizer = None
